+++ wand/api.py
@@ -294,6 +294,8 @@ try:
 
     library.MagickFlipImage.argtypes = [ctypes.c_void_p]
 
+    library.MagickFlopImage.argtypes = [ctypes.c_void_p]
+
     library.MagickResetImagePage.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
 
     library.MagickResizeImage.argtypes = [ctypes.c_void_p, ctypes.c_size_t,